Equipamento: Move referências para local correto
