+++ src/app/cadastro-equipamento/cadastro-equipamento.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Equipamento } from '../equipamento';
-import { EquipamentoService } from '../equipamento.service';
+import { Equipamento } from '../classes/equipamento';
+import { EquipamentoService } from '../services/equipamento.service';
 
 @Component({
   selector: 'app-cadastro-equipamento',
